fix tests after move to models and adding name to pallets
