+++ tests/data/list_pallets/single_pallet.py
@@ -6,11 +6,11 @@ single_pallet.py
 A module that contains pallets to be used in test_lift.py tests
 '''
 
-from forklift.pallet import Pallet
+from forklift.models import Pallet
 
 
 class SinglePallet(Pallet):
 
-    def __init__(self):
+    def __init__(self, name):
         super(SinglePallet, self).__init__()
         self.dependencies = ['a', 'b']
